Refactor OutputFormatter to remove tight coupling.
